Add option of pin button reposition (between icon & tabname)
After adding pin tab button beside of close tab button on tab, many users complain about clicking on the wrong button. This PR add an option of separating pin tab button from close tab button by putting pin button between the saved/edited icon and the tabname.

Fix #15963

+++ PowerEditor/src/Notepad_plus_Window.cpp
@@ -153,7 +153,7 @@ void Notepad_plus_Window::init(HINSTANCE hInst, HWND parent, const wchar_t *cmdL
 
 	if (cmdLineParams->_isNoTab || (nppGUI._tabStatus & TAB_HIDE))
 	{
-		const int tabStatusOld = nppGUI._tabStatus;
+		const size_t tabStatusOld = nppGUI._tabStatus;
 		::SendMessage(_hSelf, NPPM_HIDETABBAR, 0, TRUE);
 		if (cmdLineParams->_isNoTab)
 		{
